Split platform independent computation into separate module
